Move `print_executables_available`
