+++ src/cli/global/install.rs
@@ -24,7 +24,10 @@ use rattler_solve::{resolvo::Solver, SolverImpl, SolverTask};
 use rattler_virtual_packages::VirtualPackage;
 use reqwest_middleware::ClientWithMiddleware;
 
-use crate::global::{channel_name_from_prefix, find_designated_package, BinDir, BinEnvDir};
+use crate::global::{
+    channel_name_from_prefix, find_designated_package, print_executables_available, BinDir,
+    BinEnvDir,
+};
 use crate::{
     cli::cli_config::ChannelsConfig, cli::has_specs::HasSpecs, prefix::Prefix,
     rlimit::try_increase_rlimit_to_sensible,
@@ -375,34 +378,6 @@ pub async fn execute(args: Args) -> miette::Result<()> {
     Ok(())
 }
 
-async fn print_executables_available(executables: Vec<PathBuf>) -> miette::Result<()> {
-    let BinDir(bin_dir) = BinDir::from_existing().await?;
-    let whitespace = console::Emoji("  ", "").to_string();
-    let executable = executables
-        .into_iter()
-        .map(|path| {
-            path.strip_prefix(&bin_dir)
-                .expect("script paths were constructed by joining onto BinDir")
-                .to_string_lossy()
-                .to_string()
-        })
-        .join(&format!("\n{whitespace} -  "));
-
-    if is_bin_folder_on_path().await {
-        eprintln!(
-            "{whitespace}These executables are now globally available:\n{whitespace} -  {executable}",
-        )
-    } else {
-        eprintln!("{whitespace}These executables have been added to {}\n{whitespace} -  {executable}\n\n{} To use them, make sure to add {} to your PATH",
-                  console::style(&bin_dir.display()).bold(),
-                  console::style("!").yellow().bold(),
-                  console::style(&bin_dir.display()).bold()
-        )
-    }
-
-    Ok(())
-}
-
 /// Install given package globally, with all its dependencies
 pub(super) async fn globally_install_package(
     package_name: &PackageName,
@@ -483,18 +458,4 @@ fn get_catch_all_arg(shell: &ShellEnum) -> &str {
         ShellEnum::PowerShell(_) => "@args",
         _ => "\"$@\"",
     }
-}
-
-/// Returns true if the bin folder is available on the PATH.
-async fn is_bin_folder_on_path() -> bool {
-    let bin_path = match BinDir::from_existing().await.ok() {
-        Some(BinDir(bin_dir)) => bin_dir,
-        None => return false,
-    };
-
-    std::env::var_os("PATH")
-        .map(|path| std::env::split_paths(&path).collect_vec())
-        .unwrap_or_default()
-        .into_iter()
-        .contains(&bin_path)
 }
