+++ src/global/mod.rs
@@ -6,6 +6,6 @@ mod project;
 
 pub(crate) use common::{
     bin_dir, bin_env_dir, channel_name_from_prefix, find_designated_package,
-    find_installed_package, BinDir, BinEnvDir,
+    find_installed_package, print_executables_available, BinDir, BinEnvDir,
 };
 pub(crate) use project::Project;
